% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{add_component_to_pipeline}
\alias{add_component_to_pipeline}
\title{add a base component to a specified pipeline}
\usage{
add_component_to_pipeline(component, envir)
}
\arguments{
\item{component}{an object o either \code{Node} or \code{Splitter}}

\item{envir}{calling environment. Used for construction of calls from the pipeline entry}
}
\description{
add a base component to a specified pipeline
}
