% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{add_node}
\alias{add_node}
\title{add node to a sewage pipeline}
\usage{
add_node(pipeline, component, name, input, ...)
}
\arguments{
\item{pipeline}{an initialized  sewage pipeline}

\item{component}{a function to be executed. Must be a valid function specification}

\item{name}{a name to give to the given component. This will be used as the `input` parameter for downstream nodes}

\item{input}{the node to use as input into `component`. Input must already be specified in the pipeline. Inputs that create circular dependencies will throw an error.}

\item{...}{additional arguments to be passed to the `component` argument}
}
\value{
a `sewage_pipeline` object
}
\description{
`add_node()` will place a new node in the specified pipeline. This will be executed sequentially when the pipeline is executed using `run()`
}
