% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run}
\alias{run}
\title{Run a pipeline}
\usage{
run(pipeline, ...)
}
\arguments{
\item{pipeline}{an initialized pipeline object}

\item{...}{parameter to pass to first node of the pipeline. This should match the \code{input} parameter of \code{add_node} of the first node.}
}
\description{
This function is the extry point for executing a pipeline object
}
