% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewage-package.R
\docType{package}
\name{sewage-package}
\alias{sewage}
\alias{sewage-package}
\title{sewage: A light-weight pipelining tool for data analysis}
\description{
Moving from dense scripts of functions and workflows to a full-fledged ETL or DAG tool can be daunting. The gola of 'sewage' is to be a lightweight tool for organizing data analysis pipelines and make complex workflows more modular and easier to interpret.
}
\keyword{internal}
