% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Splitter}
\alias{Splitter}
\title{Initialize a splitter objects}
\usage{
Splitter(edges = 2)
}
\arguments{
\item{edges}{number out outputs. Must be greater than 1}
}
\description{
Initialize a splitter objects
}
